+++ Frontend/src/app/features/cards/services/cards.service.ts
@@ -46,8 +46,8 @@ export class CardsService extends BaseService {
         return this.httpDelete(id);
     }
 
-    public assign(id: string, userId: string): Observable<void> {
-        return this.httpPostVoid(`${id}/assign`, { userId });
+    public assign(id: string, userId: string | null): Observable<void> {
+        return this.httpPostVoid(`${id}/assign`, { userId: userId });
     }
 
     public move(id: string, newColumnId: string): Observable<void> {
@@ -56,5 +56,9 @@ export class CardsService extends BaseService {
 
     public reorder(columnId: string, orders: CardOrderDto[]): Observable<void> {
         return this.httpPostVoid(`${columnId}/reorder`, orders);
+    }
+
+    public transfer(id: string, boardId: string) {
+        return this.httpPostVoid(`${id}/transfer`, { boardId });
     }
 }