+++ Frontend/src/app/features/columns/dto/column.dto.ts
@@ -2,5 +2,6 @@ export interface ColumnDto {
     id: string;
     name: string;
     order: number;
+    locked: boolean;
     boardId: string;
 }