+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# [v2.1.0] - 2026-01-23
+
+## Added
+
+- Card transfer between boards;
+- Card assign system;
+- Locked system "New" and "Done" columns;
+- Card duplication system.
+
 # [v2.0.1] - 2026-01-23
 
 ## Fixed